+++ .gitignore
@@ -6,3 +6,4 @@ __pycache__/
 .cache/
 build/
 dist/
+whotrackme.egg-info/
