+++ whotracksme/data/__init__.py
@@ -1,0 +1,6 @@
+
+from whotracksme.data.db.trackerdb import load_tracker_db
+
+__all__ = [
+    "load_tracker_db"
+]
