+++ whotracksme/data/db/trackerdb.py
@@ -7,12 +7,12 @@ def load_tracker_db(loc=':memory:'):
     with connection:
         connection.executescript(pkgutil.get_data(
             'whotracksme',
-            'db/trackerdb.sql'
+            'data/db/trackerdb.sql'
         ).decode('utf-8'))
     return connection
 
 
-def dump_tracker_db(conn):
-    with open(_tracker_db_path(), 'w') as fp:
-        for line in conn.iterdump():
-            fp.write('%s\n' % line)
+# def dump_tracker_db(conn):
+#     with open(_tracker_db_path(), 'w') as fp:
+#         for line in conn.iterdump():
+#             fp.write('%s\n' % line)
