+++ whotracksme/website/plotting/trackers.py
@@ -2,9 +2,9 @@ from statistics import mean
 import plotly.graph_objs as go
 import squarify
 
-from whotracksme.plotting.utils import set_margins, div_output, CliqzFonts, annotation
-from whotracksme.plotting.colors import CliqzColors
-from whotracksme.plotting.plots import scatter, hbar, line
+from whotracksme.website.plotting.utils import set_margins, div_output, CliqzFonts, annotation
+from whotracksme.website.plotting.colors import CliqzColors
+from whotracksme.website.plotting.plots import scatter, hbar, line
 
 
 def tracker_cfh(https, fingerprinting, cookies):
