+++ whotracksme/website/templates.py
@@ -1,4 +1,6 @@
-from urllib.parse import quote_plus
+
+from collections import defaultdict
+import json
 import os
 import shutil
 import subprocess
@@ -6,10 +8,61 @@ import subprocess
 from jinja2 import Environment, FileSystemLoader, Markup
 import markdown
 
+from whotracksme.website.utils import print_progress
+from whotracksme.website.plotting.colors import TrackerColors, SiteCategoryColors
 
-from whotracksme.db.trackerdb import load_tracker_db
-from whotracksme.plotting.colors import TrackerColors, SiteCategoryColors
 
+def normalize(s):
+    # TODO: add other normalization features
+    return s.lower()
+
+
+def site_to_json(data_source, blog_posts):
+    site_idx = defaultdict(list)
+
+    def submit_key(name, type, url, weight, idx=site_idx):
+        _name = normalize(name)
+        idx[type].append({
+            "name": name,
+            "normalized_name": _name,
+            "type": type,
+            "url": url,
+            "weight": weight
+        })
+
+    for app_id, app in data_source.apps.items():
+        submit_key(
+            name=data_source.get_app_name(app_id),
+            type="tracker",
+            url=data_source.url_for("app", app_id),
+            weight=(1.0 / app.get("rank", 1)) * 1000
+        )
+
+    for company_id, company in data_source.companies.items():
+        submit_key(
+            name=data_source.get_company_name(company_id),
+            type="company",
+            url=data_source.url_for("company", company_id),
+            weight=len(company.get("overview", {}).get("apps", {})) or 1
+        )
+
+    for site_id, site in data_source.sites.items():
+        submit_key(
+            name=data_source.get_site_name(site_id),
+            type="site",
+            url=data_source.url_for("site", site_id),
+            weight=site.get("overview", {}).get("popularity", 0.01) * 10000
+        )
+
+    for blogpost in blog_posts:
+        submit_key(
+            name=blogpost.get("title"),
+            type="blog",
+            url=data_source.url_for("blog", blogpost.get("filename")),
+            weight=1
+        )
+
+    return site_idx
 
 # Paths needed for generating urls
 SITE_PATH = os.path.abspath('_site')
@@ -38,99 +91,28 @@ CATEGORY_DESC = {
     "extensions": "This are browser addons that collect user data"
 }
 
+def copy_custom_error_pages(data):
+    error_pages = {
+        "not-found": get_template(data, "not-found.html"),
+        "tracker-not-found": get_template(data, "tracker-not-found.html"),
+        "website-not-found": get_template(data, "website-not-found.html")
+    }
 
-class DataSource:
-    def __init__(self, overview, apps, companies, sites):
-        self.overview = overview
-        self.apps = apps
-        self.companies = companies
-        self.sites = sites
+    for error, template in error_pages.items():
+        with open('_site/{}.html'.format(error), 'w') as output:
+            output.write(render_template(template=template))
 
-        connection = load_tracker_db()
-        with connection:
-            self.app_info = self.load_app_info(connection)
-            self.company_info = self.load_company_info(connection)
 
-        for id, app in self.apps.items():
-            if id in self.app_info:
-                for k, v in self.app_info[id].items():
-                    app[k] = v
-
-        for id, company in self.companies.items():
-            if id in self.company_info:
-                for k, v in self.company_info[id].items():
-                    company[k] = v
-
-    @staticmethod
-    def normalize_url(url_substring):
-        return quote_plus(url_substring.replace("/", " ")).lower()
-
-    @staticmethod
-    def rank_label(r):
-        if r < 3:
-            return "Dangerously prevalent"
-        if 3 <= r < 11:
-            return "Extremely prevalent"
-        if 11 <= r < 50:
-            return "Very prevalent"
-        if 51 <= r < 100:
-            return "Commonly prevalent"
-        if 101 <= r:
-            return "Relatively prevalent"
-
-    def url_for(self, entity, id, path_to_root='.'):
-        if entity == "app":
-            return "{}/trackers/{}.html".format(path_to_root, self.normalize_url(id))
-        elif entity == "company":
-            return "{}/companies/{}.html".format(path_to_root, self.normalize_url(id))
-        elif entity == "site":
-            return "{}/websites/{}.html".format(path_to_root, self.get_site_name(id)).lower()
-        elif entity == "blog":
-            return "{}/blog/{}.html".format(path_to_root, id)
-
-    def get_app_name(self, id):
-        return self.app_info.get(id).get('name') if id in self.app_info else id
-
-    def get_company_name(self, id):
-        return self.company_info.get(id).get('name') if id in self.company_info else id
-
-    def get_site_name(self, id):
-        return id if id in self.sites else None
-
-    def load_app_info(self, connection):
-        col_names = ['id', 'name', 'description', 'cat', 'website_url', 'logo_url', 'company_id']
-        cur = connection.execute('''SELECT trackers.id, trackers.name, description, categories.name, website_url, logo_url,
-                company_id
-            FROM trackers
-            LEFT JOIN categories ON categories.id = category_id
-            ''')
-        app_info = {}
-        for row in cur:
-            app_info[row[0]] = {col: row[i] for i, col in enumerate(col_names)}
-            app_info[row[0]]['domains'] = []
-
-        cur = connection.execute('SELECT tracker, domain FROM tracker_domains')
-        for app, domain in cur:
-            app_info[app]['domains'].append(domain)
-
-        return app_info
-
-    def load_company_info(self, connection):
-        col_names = ['id', 'name', 'description', 'about_us_url', 'privacy_contact_url', 'privacy_url', 'website_url',
-                     'in_their_own_words', 'logo_url']
-        cur = connection.execute('''SELECT {}
-            FROM companies
-            '''.format(','.join(col_names)))
-        company_info = {}
-        for row in cur:
-            company_info[row[0]] = {col: row[i] for i, col in enumerate(col_names)}
-            company_info[row[0]]['apps'] = {}
-
-        cur = connection.execute('SELECT company_id, id, name FROM trackers WHERE company_id IS NOT NULL')
-        for cid, app_id, app_name in cur:
-            company_info[cid]['apps'][app_id] = app_name
-
-        return company_info
+def generate_sitemap(data, blog_posts):
+    # write sitemap to _site (to be used as index for static site search)
+    with open("_site/sitemap.json", "w") as output:
+        json.dump(
+            site_to_json(
+                data_source=data,
+                blog_posts=blog_posts),
+            output
+        )
+    print_progress(text='Generate sitemap index')
 
 
 def get_template(data_source, name, render_markdown=False, path_to_root='.'):
@@ -225,4 +207,3 @@ def create_site_structure(static_path):
         shutil.copytree(static_path, _site_static)
     elif not os.path.exists(_site_static):
         shutil.copytree(static_path, _site_static)
-
