+++ whotracksme/website/utils.py
@@ -1,0 +1,5 @@
+
+def print_progress(text, default_space=40):
+    print("{} {:{}} done".format(text, "." * (default_space - len(text)), default_space - len(text)))
+
+
